+++ lib/frontend/listings.c
@@ -427,22 +427,22 @@ listings_finish (lstgctx_t ctx)
 void listing_mainscope_set (lstgctx_t ctx, scopectx_t mainscope) {
     ctx->main_state.this_scope = mainscope; }
 void listing_title_set (lstgctx_t ctx, strdesc_t *str) {
-    int len = (str->len > TITLE_LEN ? TITLE_LEN : str->len);
+    size_t len = (str->len > TITLE_LEN ? TITLE_LEN : str->len);
     memcpy(ctx->header1+TITLE_POS, str->ptr, len);
     if (len < TITLE_LEN) memset(ctx->header1+(TITLE_POS+len), ' ', TITLE_LEN-len); }
 void listing_subtitle_set (lstgctx_t ctx, strdesc_t *str) {
-    int len = (str->len > SBTTL_LEN ? SBTTL_LEN : str->len);
+    size_t len = (str->len > SBTTL_LEN ? SBTTL_LEN : str->len);
     memcpy(ctx->header2+SBTTL_POS, str->ptr, len);
     if (len < SBTTL_LEN) memset(ctx->header2+(SBTTL_POS+len), ' ', SBTTL_LEN-len); }
 void listing_compilerid_set (lstgctx_t ctx, strdesc_t *str) {
-    int len = (str->len > COMPID_LEN ? COMPID_LEN : str->len);
+    size_t len = (str->len > COMPID_LEN ? COMPID_LEN : str->len);
     memcpy(ctx->header1+COMPID_POS, str->ptr, len);
     if (len < COMPID_LEN) memset(ctx->header1+(COMPID_POS+len), ' ', COMPID_LEN-len); }
 void listing_ident_set (lstgctx_t ctx, strdesc_t *str) {
-    int len = (str->len > IDENT_LEN ? IDENT_LEN : str->len);
+    size_t len = (str->len > IDENT_LEN ? IDENT_LEN : str->len);
     memcpy(ctx->header2+IDENT_POS, str->ptr, len);
     if (len < IDENT_LEN) memset(ctx->header2+(IDENT_POS+len), ' ', IDENT_LEN-len); }
 void listing_name_set (lstgctx_t ctx, strdesc_t *str) {
-    int len = (str->len > NAME_LEN ? NAME_LEN : str->len);
+    size_t len = (str->len > NAME_LEN ? NAME_LEN : str->len);
     memcpy(ctx->header1+NAME_POS, str->ptr, len);
     if (len < NAME_LEN) memset(ctx->header1+(NAME_POS+len), ' ', NAME_LEN-len); }
